+++ Packages/UGF.Logs/Runtime/Log.cs
@@ -24,7 +24,7 @@ namespace UGF.Logs.Runtime
         /// <summary>
         /// Logs message as info with the specified message and arguments.
         /// <para>
-        /// Invocation of this method will be included in release build, only if the `UGF_LOG_INFO` compilation symbol is defined.
+        /// Invocation of this method will be included in release build, only if the 'UGF_LOG_INFO' compilation symbol is defined.
         /// </para>
         /// </summary>
         /// <param name="message">The formattable message.</param>
@@ -32,8 +32,12 @@ namespace UGF.Logs.Runtime
         /// <param name="arg1">The second argument.</param>
         /// <param name="arg2">The third argument.</param>
         [Conditional("UGF_LOG_INFO")]
+#if UGF_LOG_INCLUDE_EDITOR
         [Conditional("UNITY_EDITOR")]
+#endif
+#if UGF_LOG_INCLUDE_DEVBUILD
         [Conditional("DEVELOPMENT_BUILD")]
+#endif
         [StringFormatMethod("message")]
         public static void Info(string message, object arg0 = null, object arg1 = null, object arg2 = null)
         {
@@ -43,14 +47,18 @@ namespace UGF.Logs.Runtime
         /// <summary>
         /// Logs message as info with the specified message and arguments.
         /// <para>
-        /// Invocation of this method will be included in release build, only if the `UGF_LOG_INFO` compilation symbol is defined.
+        /// Invocation of this method will be included in release build, only if the 'UGF_LOG_INFO' compilation symbol is defined.
         /// </para>
         /// </summary>
         /// <param name="message">The formattable message.</param>
         /// <param name="args">The arguments.</param>
         [Conditional("UGF_LOG_INFO")]
+#if UGF_LOG_INCLUDE_EDITOR
         [Conditional("UNITY_EDITOR")]
+#endif
+#if UGF_LOG_INCLUDE_DEVBUILD
         [Conditional("DEVELOPMENT_BUILD")]
+#endif
         [StringFormatMethod("message")]
         public static void Info(string message, params object[] args)
         {
@@ -60,7 +68,7 @@ namespace UGF.Logs.Runtime
         /// <summary>
         /// Logs message as debug info with the specified message and arguments.
         /// <para>
-        /// Invocation of this method will be included in release build, only if the `UGF_LOG_DEBUG` compilation symbol is defined.
+        /// Invocation of this method will be included in release build, only if the 'UGF_LOG_DEBUG' compilation symbol is defined.
         /// </para>
         /// </summary>
         /// <param name="message">The formattable message.</param>
@@ -68,8 +76,12 @@ namespace UGF.Logs.Runtime
         /// <param name="arg1">The second argument.</param>
         /// <param name="arg2">The third argument.</param>
         [Conditional("UGF_LOG_DEBUG")]
+#if UGF_LOG_INCLUDE_EDITOR
         [Conditional("UNITY_EDITOR")]
+#endif
+#if UGF_LOG_INCLUDE_DEVBUILD
         [Conditional("DEVELOPMENT_BUILD")]
+#endif
         [StringFormatMethod("message")]
         public static void Debug(string message, object arg0 = null, object arg1 = null, object arg2 = null)
         {
@@ -79,14 +91,18 @@ namespace UGF.Logs.Runtime
         /// <summary>
         /// Logs message as debug info with the specified message and arguments.
         /// <para>
-        /// Invocation of this method will be included in release build, only if the `UGF_LOG_INFO` compilation symbol is defined.
+        /// Invocation of this method will be included in release build, only if the 'UGF_LOG_INFO' compilation symbol is defined.
         /// </para>
         /// </summary>
         /// <param name="message">The formattable message.</param>
         /// <param name="args">The arguments.</param>
         [Conditional("UGF_LOG_DEBUG")]
+#if UGF_LOG_INCLUDE_EDITOR
         [Conditional("UNITY_EDITOR")]
+#endif
+#if UGF_LOG_INCLUDE_DEVBUILD
         [Conditional("DEVELOPMENT_BUILD")]
+#endif
         [StringFormatMethod("message")]
         public static void Debug(string message, params object[] args)
         {
@@ -96,7 +112,7 @@ namespace UGF.Logs.Runtime
         /// <summary>
         /// Logs message as warning with the specified tag, message and arguments.
         /// <para>
-        /// Invocation of this method will be included in release build, only if the `UGF_LOG_WARNING` compilation symbol is defined.
+        /// Invocation of this method will be included in release build, only if the 'UGF_LOG_WARNING' compilation symbol is defined.
         /// </para>
         /// </summary>
         /// <param name="message">The formattable message.</param>
@@ -104,8 +120,12 @@ namespace UGF.Logs.Runtime
         /// <param name="arg1">The second argument.</param>
         /// <param name="arg2">The third argument.</param>
         [Conditional("UGF_LOG_WARNING")]
+#if UGF_LOG_INCLUDE_EDITOR
         [Conditional("UNITY_EDITOR")]
+#endif
+#if UGF_LOG_INCLUDE_DEVBUILD
         [Conditional("DEVELOPMENT_BUILD")]
+#endif
         [StringFormatMethod("message")]
         public static void Warning(string message, object arg0 = null, object arg1 = null, object arg2 = null)
         {
@@ -115,14 +135,18 @@ namespace UGF.Logs.Runtime
         /// <summary>
         /// Logs message as warning with the specified tag, message and arguments.
         /// <para>
-        /// Invocation of this method will be included in release build, only if the `UGF_LOG_WARNING` compilation symbol is defined.
+        /// Invocation of this method will be included in release build, only if the 'UGF_LOG_WARNING' compilation symbol is defined.
         /// </para>
         /// </summary>
         /// <param name="message">The formattable message.</param>
         /// <param name="args">The arguments.</param>
         [Conditional("UGF_LOG_WARNING")]
+#if UGF_LOG_INCLUDE_EDITOR
         [Conditional("UNITY_EDITOR")]
+#endif
+#if UGF_LOG_INCLUDE_DEVBUILD
         [Conditional("DEVELOPMENT_BUILD")]
+#endif
         [StringFormatMethod("message")]
         public static void Warning(string message, params object[] args)
         {
@@ -132,7 +156,7 @@ namespace UGF.Logs.Runtime
         /// <summary>
         /// Logs message as error with the specified tag, message and arguments.
         /// <para>
-        /// Invocation of this method will be included in release build, only if the `UGF_LOG_ERROR` compilation symbol is defined.
+        /// Invocation of this method will be included in release build, only if the 'UGF_LOG_ERROR' compilation symbol is defined.
         /// </para>
         /// </summary>
         /// <param name="message">The formattable message.</param>
@@ -140,8 +164,12 @@ namespace UGF.Logs.Runtime
         /// <param name="arg1">The second argument.</param>
         /// <param name="arg2">The third argument.</param>
         [Conditional("UGF_LOG_ERROR")]
+#if UGF_LOG_INCLUDE_EDITOR
         [Conditional("UNITY_EDITOR")]
+#endif
+#if UGF_LOG_INCLUDE_DEVBUILD
         [Conditional("DEVELOPMENT_BUILD")]
+#endif
         [StringFormatMethod("message")]
         public static void Error(string message, object arg0 = null, object arg1 = null, object arg2 = null)
         {
@@ -151,14 +179,18 @@ namespace UGF.Logs.Runtime
         /// <summary>
         /// Logs message as error with the specified tag, message and arguments.
         /// <para>
-        /// Invocation of this method will be included in release build, only if the `UGF_LOG_ERROR` compilation symbol is defined.
+        /// Invocation of this method will be included in release build, only if the 'UGF_LOG_ERROR' compilation symbol is defined.
         /// </para>
         /// </summary>
         /// <param name="message">The formattable message.</param>
         /// <param name="args">The arguments.</param>
         [Conditional("UGF_LOG_ERROR")]
+#if UGF_LOG_INCLUDE_EDITOR
         [Conditional("UNITY_EDITOR")]
+#endif
+#if UGF_LOG_INCLUDE_DEVBUILD
         [Conditional("DEVELOPMENT_BUILD")]
+#endif
         [StringFormatMethod("message")]
         public static void Error(string message, params object[] args)
         {
@@ -168,13 +200,17 @@ namespace UGF.Logs.Runtime
         /// <summary>
         /// Logs message based on the specified exception.
         /// <para>
-        /// Invocation of this method will be included in release build, only if the `UGF_LOG_EXCEPTION` compilation symbol is defined.
+        /// Invocation of this method will be included in release build, only if the 'UGF_LOG_EXCEPTION' compilation symbol is defined.
         /// </para>
         /// </summary>
         /// <param name="exception">The exception to log.</param>
         [Conditional("UGF_LOG_EXCEPTION")]
+#if UGF_LOG_INCLUDE_EDITOR
         [Conditional("UNITY_EDITOR")]
+#endif
+#if UGF_LOG_INCLUDE_DEVBUILD
         [Conditional("DEVELOPMENT_BUILD")]
+#endif
         public static void Exception(Exception exception)
         {
             if (exception == null) throw new ArgumentNullException(nameof(exception));
@@ -185,7 +221,7 @@ namespace UGF.Logs.Runtime
         /// <summary>
         /// Logs message with the specified log type, message and arguments.
         /// <para>
-        /// Invocation of this method always include in release build.
+        /// Invocation of this method always included in release build.
         /// </para>
         /// </summary>
         /// <param name="logType">The type of the log.</param>
@@ -210,7 +246,7 @@ namespace UGF.Logs.Runtime
         /// <summary>
         /// Logs message with the specified log type, message and arguments.
         /// <para>
-        /// Invocation of this method always include in release build.
+        /// Invocation of this method always included in release build.
         /// </para>
         /// </summary>
         /// <param name="logType">The type of the log.</param>